pyproject.toml
[project]
name = "rookie-ranker"
version = "0.1.0"
description = "Predict and rank NFL rookies for fantasy football drafts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.3",
    "nflreadpy>=0.1.5",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.0",
    "requests>=2.31",
    "scikit-learn>=1.4",
    "shap>=0.44",
]
